tests for long, float, integer and double division added literal parsing for new types except date added fixed minor bugs

+++ jshuntingyard/src/main/java/org/oss/jshuntingyard/evaluator/FloatArgument.java
@@ -24,7 +24,7 @@ public class FloatArgument extends AbstractFunctionElementArgument<Float> implem
 
 	@Override
 	public FunctionElementArgument.ArgumentType getType() {
-		return FunctionElementArgument.ArgumentType.DOUBLE;
+		return FunctionElementArgument.ArgumentType.FLOAT;
 	}
 
 	@Override
